user-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
requests
PyJWT
python-multipart
python-jose[cryptography]
httpx
prometheus-client
